Minor refactoring and docs fixes

+++ dask-gateway-server/setup.py
@@ -6,7 +6,6 @@ of this, this setup.py file includes some related complexity.
 
 Relevant reference documentation:
 - https://setuptools.pypa.io/en/latest/userguide/extension.html#adding-commands
-
 """
 import errno
 import os
@@ -26,12 +25,6 @@ ROOT_DIR = os.path.abspath(os.path.dirname(__file__))
 PROXY_SRC_DIR = os.path.join(ROOT_DIR, "dask-gateway-proxy")
 PROXY_TGT_DIR = os.path.join(ROOT_DIR, "dask_gateway_server", "proxy")
 PROXY_TGT_EXE = os.path.join(PROXY_TGT_DIR, "dask-gateway-proxy")
-
-# determine version from _version.py
-ns = {}
-with open(os.path.join(ROOT_DIR, "dask_gateway_server", "_version.py")) as f:
-    exec(f.read(), {}, ns)
-    VERSION = ns["__version__"]
 
 
 class build_proxy(Command):
@@ -77,7 +70,8 @@ class build_proxy(Command):
 class build_proxy_mixin:
     """
     This mixin class helps us ensure that we build the dask-gateway-proxy
-    executable binary independent of if we run the install, build, develop.
+    executable binary no matter if we run the: install, build, develop, or
+    bdist_wheel command.
     """
 
     def run(self):
@@ -88,19 +82,6 @@ class build_proxy_mixin:
         ):
             self.run_command("build_proxy")
         super().run()
-
-
-if os.environ.get("DASK_GATEWAY_SERVER__NO_PROXY"):
-    package_data = {}
-else:
-    # FIXME: Does it make sense to put the binary next to the source code
-    #        (proxy/dask-gateway-proxy)? If we build a wheel now we get a folder in
-    #        the wheel called purelib that includes this and it seems contradictory.
-    #
-    #        An action point to resolve this would be to investigate how other
-    #        projects have bundled compiled binaries.
-    #
-    package_data = {"dask_gateway_server": ["proxy/dask-gateway-proxy"]}
 
 
 class build(build_proxy_mixin, _build):
@@ -117,10 +98,9 @@ class develop(build_proxy_mixin, _develop):
 
 class clean(_clean):
     def run(self):
-        if self.all:
-            if os.path.exists(PROXY_TGT_EXE):
-                os.remove(PROXY_TGT_EXE)
-        _clean.run(self)
+        super().run()
+        if self.all and os.path.exists(PROXY_TGT_EXE):
+            os.remove(PROXY_TGT_EXE)
 
 
 try:
@@ -157,15 +137,14 @@ try:
             if os.environ.get("DASK_GATEWAY_SERVER__NO_PROXY"):
                 return python, abi, "any"
 
-            # If GOOS or GOARCH are set, we are probably cross compiling. For
-            # some limited know configurations, lets support building the wheel
-            # with cross compiling.
+            # If GOOS or GOARCH are set, we will probably be cross compiling.
+            # Below we act intelligently based on a few combinations of GOOS and
+            # GOARCH to provide a platform tag that is accepted by PyPI.
             #
             # PyPI restricts us to using the following platform tags:
             # https://github.com/pypa/warehouse/blob/82815b06d9f98deed5f205c66e054de59d22a10d/warehouse/forklift/legacy.py#L108-L180
             #
-            # For reference, GitHub Actions provides workers with the following
-            # operating systems:
+            # For reference, GitHub Actions available operating systems:
             # https://github.com/actions/virtual-environments#available-environments
             #
             go_plat = os.environ.get("GOOS", "") + "_" + os.environ.get("GOARCH", "")
@@ -183,14 +162,44 @@ try:
 except ImportError:
     bdist_wheel = None
 
+# cmdclass describes commands we can run with "python setup.py <command>". We
+# have overridden several command classes by mixing in a common capability of
+# compiling the golang executable binary.
+#
+# It seems that only by overriding multiple command classes we can ensure we
+# manage to compile the Golang executable binary before its needed. Below are
+# examples of end user commands that are impacted by our changes to cmdclass.
+#
+#   python setup.py [bdist_wheel|build|install|develop|clean]
+#   python -m pip install [--editable] .
+#   python -m build --wheel .
+#
 cmdclass = {
-    "bdist_wheel": bdist_wheel,  # update built wheel descriptions
-    "build_proxy": build_proxy,  # directly build the proxy source
-    "build": build,  # bdist_wheel or pip install .
-    "install": install,  # python setup.py install
-    "develop": develop,  # python setup.py develop
+    "bdist_wheel": bdist_wheel,
+    "build": build,
+    "build_proxy": build_proxy,
+    "install": install,
     "clean": clean,
+    "develop": develop,
 }
+
+if os.environ.get("DASK_GATEWAY_SERVER__NO_PROXY"):
+    package_data = {}
+else:
+    # FIXME: Does it make sense to put the binary next to the source code
+    #        (proxy/dask-gateway-proxy)? If we build a wheel now we get a folder in
+    #        the wheel called purelib that includes this and it seems contradictory.
+    #
+    #        An action point to resolve this would be to investigate how other
+    #        projects have bundled compiled binaries.
+    #
+    package_data = {"dask_gateway_server": ["proxy/dask-gateway-proxy"]}
+
+# determine version from _version.py
+ns = {}
+with open(os.path.join(ROOT_DIR, "dask_gateway_server", "_version.py")) as f:
+    exec(f.read(), {}, ns)
+    VERSION = ns["__version__"]
 
 with open("requirements.txt") as f:
     install_requires = [l for l in f.readlines() if not l.startswith("#")]
